ScanOperator created #149
