+++ libakumuli/storage_engine/operators/scan.h
@@ -1,0 +1,41 @@
+#pragma once
+
+#include "operator.h"
+
+namespace Akumuli {
+namespace StorageEngine {
+
+
+// ////////////////////////////// //
+//  NBTreeIterator concatenation  //
+// ////////////////////////////// //
+
+/** Concatenating iterator.
+  * Accepts list of iterators in the c-tor. All iterators then
+  * can be seen as one iterator. Iterators should be in correct
+  * order.
+  */
+struct ScanOperator : RealValuedOperator {
+    typedef std::vector<std::unique_ptr<RealValuedOperator>> IterVec;
+    IterVec   iter_;
+    Direction dir_;
+    u32       iter_index_;
+
+    //! C-tor. Create iterator from list of iterators.
+    template<class TVec>
+    ScanOperator(TVec&& iter)
+        : iter_(std::forward<TVec>(iter))
+        , iter_index_(0)
+    {
+        if (iter_.empty()) {
+            dir_ = Direction::FORWARD;
+        } else {
+            dir_ = iter_.front()->get_direction();
+        }
+    }
+
+    virtual std::tuple<aku_Status, size_t> read(aku_Timestamp *destts, double *destval, size_t size);
+    virtual Direction get_direction();
+};
+
+}}  // namespace
